* fixed faucet delay ui error.

* nit fixes.

* updated automated gosdk replacement gradle tasks.

* added gradient for icons in SelectAppFragment

* updated gosdk to latest PR #971.

+++ app/src/main/java/org/zus/helloworld/models/vult/AllocationModelItem.kt
@@ -8,7 +8,7 @@ data class AllocationModelItem(
     @SerializedName("data_shards")
     val dataShards: Int,
     @SerializedName("expiration_date")
-    val expirationDate: Int,
+    val expirationDate: Long,
     @SerializedName("id")
     val id: String,
     @SerializedName("name")
